+++ packages/plugin/src/components/ApiItem.tsx
@@ -1,10 +1,10 @@
-import React, { useMemo } from 'react';
-import type { JSONOutput } from 'typedoc';
+import { useMemo } from 'react';
 import { PageMetadata } from '@docusaurus/theme-common';
 import type { Props as DocItemProps } from '@theme/DocItem';
-import { useReflection } from '../hooks/useReflection';
+import { useReflection, useRequiredReflection } from '../hooks/useReflection';
 import { useReflectionMap } from '../hooks/useReflectionMap';
-import type { DeclarationReflectionMap, TOCItem } from '../types';
+import type { TOCItem, TSDDeclarationReflection, TSDDeclarationReflectionMap } from '../types';
+import { escapeMdx } from '../utils/helpers';
 import { getKindIconHtml } from '../utils/icons';
 import ApiItemLayout from './ApiItemLayout';
 import { displayPartsToMarkdown } from './Comment';
@@ -12,29 +12,27 @@ import { Flags } from './Flags';
 import { Reflection } from './Reflection';
 import { TypeParametersGeneric } from './TypeParametersGeneric';
 
-function extractTOC(
-	item: JSONOutput.DeclarationReflection,
-	map: DeclarationReflectionMap,
-): TOCItem[] {
+function extractTOC(item: TSDDeclarationReflection, map: TSDDeclarationReflectionMap): TOCItem[] {
 	const toc: TOCItem[] = [];
 	const mapped = new Set<string>();
 
 	item.groups?.forEach((group) => {
 		group.children?.forEach((childId) => {
-			const child = map[childId]!;
+			const child = map[childId];
 
 			if (mapped.has(child.name)) {
 				return;
 			}
 
-			const iconHtml = getKindIconHtml(child.kind, child.name);
-
 			if (!child.permalink || child.permalink.includes('#')) {
+				const iconHtml = getKindIconHtml(child.kind, child.name);
+				const value = escapeMdx(child.name) ?? '';
+
 				toc.push({
 					// @ts-expect-error Not typed upstream
 					children: [],
 					id: child.name,
-					value: iconHtml ? `${iconHtml} ${child.name}` : child.name,
+					value: iconHtml ? `${iconHtml} ${value}` : value,
 					level: 1,
 				});
 
@@ -51,7 +49,7 @@ export interface ApiItemProps extends Pick<DocItemProps, 'route'> {
 }
 
 export default function ApiItem({ readme: Readme, route }: ApiItemProps) {
-	const item = useReflection((route as unknown as { id: number }).id)!;
+	const item = useRequiredReflection((route as unknown as { id: number }).id);
 	const reflections = useReflectionMap();
 	const toc = useMemo(() => extractTOC(item, reflections), [item, reflections]);
 
@@ -63,14 +61,14 @@ export default function ApiItem({ readme: Readme, route }: ApiItemProps) {
 			next: nextItem
 				? {
 						permalink: nextItem.permalink,
-						title: nextItem.name,
-				  }
+						title: escapeMdx(nextItem.name) ?? '',
+					}
 				: undefined,
 			previous: prevItem
 				? {
 						permalink: prevItem.permalink,
-						title: prevItem.name,
-				  }
+						title: escapeMdx(prevItem.name) ?? '',
+					}
 				: undefined,
 		}),
 		[nextItem, prevItem],
@@ -83,15 +81,15 @@ export default function ApiItem({ readme: Readme, route }: ApiItemProps) {
 					<span className="tsd-header-flags">
 						<Flags flags={item.flags} />
 					</span>
-					{item.name} <TypeParametersGeneric params={item.typeParameters} />
+					{escapeMdx(item.name)} <TypeParametersGeneric params={item.typeParameters} />
 				</>
 			}
 			module={(item as unknown as { module: string }).module}
 			name={item.name}
 			pageMetadata={
 				<PageMetadata
-					description={item.comment?.summary ? displayPartsToMarkdown(item.comment.summary) : ''}
-					title={`${item.name} | API`}
+					description={item?.comment?.summary ? displayPartsToMarkdown(item.comment.summary) : ''}
+					title={`${item?.name ? `${item.name} | ` : ''}API`}
 				/>
 			}
 			pagingMetadata={pagingMetadata}