+++ packages/plugin/src/components/MemberDeclaration.tsx
@@ -1,14 +1,14 @@
 // https://github.com/TypeStrong/typedoc-default-themes/blob/master/src/default/partials/member.declaration.hbs
 
-import React from 'react';
 import { useMinimalLayout } from '../hooks/useMinimalLayout';
-import { useReflection } from '../hooks/useReflection';
+import { useRequiredReflection } from '../hooks/useReflection';
+import { escapeMdx } from '../utils/helpers';
 import { Comment, hasComment } from './Comment';
 import { DefaultValue } from './DefaultValue';
 import { Icon } from './Icon';
 import { MemberSources } from './MemberSources';
 import { Parameter } from './Parameter';
-import { Type } from './Type';
+import { extractDeclarationFromType, Type } from './Type';
 import { TypeParameters } from './TypeParameters';
 import { TypeParametersGeneric } from './TypeParametersGeneric';
 
@@ -17,17 +17,17 @@ export interface MemberDeclarationProps {
 }
 
 export function MemberDeclaration({ id }: MemberDeclarationProps) {
-	const reflection = useReflection(id)!;
+	const reflection = useRequiredReflection(id);
 	const minimal = useMinimalLayout();
 	const showTypes = reflection.typeParameters && reflection.typeParameters.length > 0;
-	const showDeclaration = !minimal && reflection.type?.declaration;
+	const showDeclaration = !minimal && extractDeclarationFromType(reflection.type);
 
 	return (
 		<>
 			<div className="tsd-panel-content">
 				<div className="tsd-signature tsd-kind-icon">
 					<Icon reflection={reflection} />
-					{reflection.name}
+					{escapeMdx(reflection.name)}
 					<TypeParametersGeneric params={reflection.typeParameters} />
 					<span className="tsd-signature-symbol">{reflection.flags?.isOptional && '?'}: </span>{' '}
 					<Type type={reflection.type} />
@@ -58,7 +58,7 @@ export function MemberDeclaration({ id }: MemberDeclarationProps) {
 				{showDeclaration && (
 					<div className="tsd-type-declaration">
 						<h4>Type declaration</h4>
-						<Parameter param={reflection.type?.declaration} />
+						<Parameter param={extractDeclarationFromType(reflection.type)} />
 					</div>
 				)}
 			</div>