+++ packages/plugin/src/utils/helpers.ts
@@ -1,0 +1,4 @@
+export function escapeMdx(value: string | null): string | null {
+	// New values are unicode!
+	return value?.replace('<', '＜').replace('>', '＞') ?? null;
+}