+++ website/docusaurus.config.ts
@@ -1,12 +1,17 @@
-const path = require('path');
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+/* eslint-disable sort-keys */
 
-let versions = [];
+import path from 'path';
+import { themes } from 'prism-react-renderer';
+import type { Config } from '@docusaurus/types';
+
+let versions: string[] = [];
 
 try {
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 	versions = require('./versions.json');
-} catch {}
+} catch {
+	// Ignore
+}
 
 // MONOREPO
 const monorepo = {
@@ -32,9 +37,14 @@ const monorepoOnePackage = {
 	packages: ['standard'],
 };
 
+const monorepoReexports = {
+	projectRoot: path.join(__dirname, '../fixtures/monorepo-1-package'),
+	packages: ['reexports'],
+};
+
 // POLYREPO STANDARD
 const polyrepo = {
-	projectRoot: path.join(__dirname, '../fixtures/polyrepo-standard'),
+	projectRoot: path.join(__dirname, '../fixtures/polyrepo'),
 	packages: ['.'],
 };
 
@@ -117,6 +127,8 @@ function getPluginConfig() {
 			return monorepo;
 		case 'monorepo-1':
 			return monorepoOnePackage;
+		case 'monorepo-reexports':
+			return monorepoReexports;
 		case 'polyrepo':
 			return polyrepo;
 		case 'polyrepo-deep':
@@ -128,8 +140,7 @@ function getPluginConfig() {
 	}
 }
 
-/** @type {import('@docusaurus/types').DocusaurusConfig} */
-module.exports = {
+const config: Config = {
 	title: 'My Site',
 	tagline: 'Dinosaurs are cool',
 	url: 'https://your-docusaurus-test-site.com',
@@ -174,7 +185,7 @@ module.exports = {
 										label: version,
 										to: i === 0 ? 'api' : `api/${version}`,
 									})),
-							  ]
+								]
 							: [],
 				},
 				{
@@ -226,8 +237,11 @@ module.exports = {
 			copyright: `Copyright © ${new Date().getFullYear()} My Project, Inc. Built with Docusaurus.`,
 		},
 		prism: {
-			theme: lightCodeTheme,
-			darkTheme: darkCodeTheme,
+			theme: themes.github,
+			darkTheme: themes.dracula,
+			prism: {
+				additionalLanguages: ['bash', 'diff', 'json', 'typescript'],
+			},
 		},
 	},
 	presets: [
@@ -264,3 +278,5 @@ module.exports = {
 		],
 	],
 };
+
+export default config;